improve: migrate hover.js to typescript
