+++ .eslintrc.json
@@ -7,7 +7,10 @@
     "ecmaVersion": 6,
     "sourceType": "module"
   },
-  "extends": "eslint:recommended",
+  "extends": [
+    "eslint:recommended",
+    "@typescript-eslint/no-unused-vars"
+  ],
   "env": {
     "browser": false,
     "commonjs": true,
